api/requirements.txt
fastapi==0.104.1
pymongo==4.6.1
python-dotenv==1.0.0
uvicorn>=0.15.0